requirements.txt
flask==3.0.*
openai==1.47.*
azure-identity==1.17.*
python-dotenv==1.0.*
gunicorn==23.0.*